mirrors: fix lumito.net mirror name in Makefile
Was renamed in commit a52ea114cea7 ("mirrors: change cdn.lumito.net to
termux.cdn.lumito.net"), but Makefile was forgotten.

It is unclear to me why the make install step continues even though an
error occurs.  There is some room for improvements here to make it
harder to do this type of mistake.